requirements.txt
Django==3.2.12
djangorestframework==3.13.1
drf-yasg==1.20.0
gunicorn==20.1.0
mysqlclient==2.1.0
urllib3==1.26.18
python-dotenv==1.0.0
django-cors-headers==3.7.0
openai==0.28.1
langchain==0.0.309
django-storages==1.13.2
PyJWT==2.6.0
djangorestframework-simplejwt==5.2.2
openpyxl==3.1.2
pdfminer.six==20201018
python-docx==0.8.11
python-pptx==0.6.21
xlrd==2.0.1
pinecone-client==2.2.1
olefile==0.46
pymongo==4.3.3
bcrypt==4.0.1
PyJWT==2.6.0
PyMuPDF==1.22.3
pypdf==3.9.1
plotly==5.15.0
django-filter==23.2
tabulate==0.9.0
bs4==0.0.1
pandas==1.5.3 
loguru==0.6.0
numpy==1.25.2
tiktoken==0.4.0
html2text==2020.1.16